Renamed All Schema files to match file naming patterns
