+++ apps/augury-backend/src/config/schemas/StockSchema.ts
